Move index.js files out of unnecessary directories.
