+++ src/pages/annual-report-2018.js
@@ -1,0 +1,9 @@
+import React from "react";
+import AnnualReportView from "../components/pages/annual-report-2018/AnnualReportView";
+import Layout from "../layouts/Layout";
+
+export default () => (
+  <Layout>
+    <AnnualReportView />
+  </Layout>
+);
