--- src/pages/annual-report-2018/index.js
@@ -1,9 +1,0 @@
-import React from "react";
-import AnnualReportView from "../../components/pages/annual-report-2018/AnnualReportView";
-import Layout from "../../layouts/Layout";
-
-export default () => (
-  <Layout>
-    <AnnualReportView />
-  </Layout>
-);
