+++ src/pages/donate.js
@@ -1,0 +1,9 @@
+import React from "react";
+import DonateView from "../components/pages/donate/DonateView";
+import Layout from "../layouts/Layout";
+
+export default () => (
+  <Layout>
+    <DonateView />
+  </Layout>
+);
