--- src/pages/donate/index.js
@@ -1,9 +1,0 @@
-import React from "react";
-import DonateView from "../../components/pages/donate/DonateView";
-import Layout from "../../layouts/Layout";
-
-export default () => (
-  <Layout>
-    <DonateView />
-  </Layout>
-);
