+++ src/pages/givingtuesday.js
@@ -1,0 +1,9 @@
+import React from "react";
+import GivingTuesdayView from "../components/pages/givingtuesday/GivingTuesdayView";
+import Layout from "../layouts/Layout";
+
+export default () => (
+  <Layout>
+    <GivingTuesdayView />
+  </Layout>
+);
