--- src/pages/givingtuesday/index.js
@@ -1,9 +1,0 @@
-import React from "react";
-import GivingTuesdayView from "../../components/pages/givingtuesday/GivingTuesdayView";
-import Layout from "../../layouts/Layout";
-
-export default () => (
-  <Layout>
-    <GivingTuesdayView />
-  </Layout>
-);
