+++ src/pages/team.js
@@ -1,0 +1,9 @@
+import React from "react";
+import TeamView from "../components/pages/team/TeamView";
+import Layout from "../layouts/Layout";
+
+export default () => (
+  <Layout>
+    <TeamView />
+  </Layout>
+);
