--- src/pages/team/index.js
@@ -1,9 +1,0 @@
-import React from "react";
-import TeamView from "../../components/pages/team/TeamView";
-import Layout from "../../layouts/Layout";
-
-export default () => (
-  <Layout>
-    <TeamView />
-  </Layout>
-);
